Add: creata logica per indice ordinamento squadra: le squadre vengono ordinate in base alla competenza e allo stato

+++ src/backend/SO115App.Models/Classi/Condivise/Squadra.cs
@@ -104,7 +104,7 @@ namespace SO115App.API.Models.Classi.Condivise
 
         public List<string> ListaCodiciFiscaliComponentiSquadra { get; set; }
 
-        public int IndiceOrdinamento { get; set; }
+        public decimal IndiceOrdinamento { get; set; }
 
         public string Turno { get; set; }
     }
